backend/requirements.txt
fastapi
uvicorn
supabase
openai
python-multipart
pydantic
python-dotenv
nltk
spacy
